removes old popup message listener
